Move ValidateTransaction function to models/transaction.go file and likewise for its unit tests
